refactor: extract TranscriptionRequest.swift from TranscriptionModels.swift
Splits the `TranscriptionRequest` Sendable value out of
TranscriptionModels.swift so that file holds only TranscriptionResult
+ TranscriptionSegment.

Byte-identical move of lines 3-20. Both files import Foundation.

TranscriptionModels.swift: 42 → 23 lines. Tests: 667.
